commit-jw

+++ Research/jeongwoon/detect.py
@@ -133,8 +133,10 @@ def detect(save_img=False):
 
                         fourcc = 'mp4v'  # output video codec
                         fps = vid_cap.get(cv2.CAP_PROP_FPS)
-                        w = int(vid_cap.get(cv2.CAP_PROP_FRAME_WIDTH))
-                        h = int(vid_cap.get(cv2.CAP_PROP_FRAME_HEIGHT))
+                        #w = float(vid_cap.get(cv2.CAP_PROP_FRAME_WIDTH))
+                        #h = float(vid_cap.get(cv2.CAP_PROP_FRAME_HEIGHT))
+                        w = float(vid_cap.get(cv2.CAP_PROP_FRAME_WIDTH))
+                        h = float(vid_cap.get(cv2.CAP_PROP_FRAME_HEIGHT))
                         vid_writer = cv2.VideoWriter(save_path, cv2.VideoWriter_fourcc(*fourcc), fps, (w, h))
                     vid_writer.write(im0)
